Updates on several sections
1) Registry Lock Contact

A registry lock contact is a contact connected with a domain object that is able to authorize changes to the object.


Since we plan to allow multiple contacts I think it would be good to rephrase the paragraph, e.G. something like:


A registry lock contact is a contact connected with a domain object that is able to authorize changes to the object.

Multiple registry lock contacts may exist for a domain. 

In this case changes may need to be authorized by more than one registry lock contact.



2) Status Values for Locked Domains

Once a registry lock has been applied on a domain object the object MUST have the serverDeleteProhibited status value set on it.


I think we should add the other status as well, maybe like


Once a registry lock has been applied on a domain object the object MUST have at least the serverDeleteProhibited status value set on it.

In addition the domain object MAY has a serverTransferProhibited and / or serverUpdateProhibited status value set on it.



3) EPP <transfer> Command
Since we plan to prohibit a transfer / the lock needs to be removed before a transfer would that not require a serverTransferProhibited in any case (MUST instead of MAY)


4) EPP <info> Command
Apologies if the question sounds silly but at .de we do not use EPP - would we not need an additional element that shows that a domain is locked?
So for the scenario: A registrar wants to transfer a domain into his account. This domain is locked, hence serverTransferProhibited is shown in RRI-INFO. 
But would it not be helpful for this registrar to also know why this status is set? e.G. by something like

<regLock:infData
S:       xmlns:regLock="urn:ietf:params:xml:ns:regLock-1.0">
S:        <regLock:policyData>
S:          <regLock:status>true</regLock:status>
S:        </regLock:policyData>
S:      </regLock:infData>


BTW, there is a typo in the Closing-Tag (secDNS instead of regLock):

S:      </secDNS:infData>



5) I think it would make sense to have a more description in the regLock:contact Element for Methods, so maybe instead of 

    An OPTIONAL <regLock:method> the method used by the registry lock contact to authorize changes

Something like

    MUST <regLock:method> the method(s) used by the registry lock contact to authorize changes. This can be one of the following methods:
        email
        text message (not limited to classic text message but also including messages to messenger services)
        letter
        phone
        totp

      Multiple methods can be set for a registry lock contact. 

BTW: Should a method not be MUST in all cases? And probably it makes sense to add an option for multiple methods, e.G. in case of a quorom one contact may prefer email while the other prefers something else